// src/domains/3d-asset-exporter/constants/asset-exporter-api.ts
export enum AssetExporterApiRoute {
  Assets = '/api/assets',
  Trigger3d = '/api/trigger-3d',
  Trigger3dStatus = '/api/trigger-3d/status',
  Trigger3dRemesh = '/api/trigger-3d/remesh',
  TriggerUpload = '/api/trigger-upload',
  SaveImage = '/api/save-image',
}

export enum MeshyApiRoute {
  ImageTo3dV1 = 'https://api.meshy.ai/v1/image-to-3d',
}